+++ src/OpenSSL/SSL.py
@@ -1529,6 +1529,10 @@ class Connection(object):
 
         ssl = _lib.SSL_new(context._context)
         self._ssl = _ffi.gc(ssl, _lib.SSL_free)
+        # We set SSL_MODE_AUTO_RETRY to handle situations where OpenSSL returns
+        # an SSL_ERROR_WANT_READ when processing a non-application data packet
+        # instead of the expected application data.
+        # See https://github.com/openssl/openssl/issues/6234 for more details.
         _lib.SSL_set_mode(self._ssl, _lib.SSL_MODE_AUTO_RETRY)
         self._context = context
         self._app_data = None